+++ cmd/login/wc_test.go
@@ -13,6 +13,8 @@ import (
 	"strings"
 	"testing"
 
+	v1 "k8s.io/api/authorization/v1"
+
 	corev1alpha1 "github.com/giantswarm/apiextensions/v6/pkg/apis/core/v1alpha1"
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v6/pkg/apis/infrastructure/v1alpha3"
 	securityv1alpha1 "github.com/giantswarm/apiextensions/v6/pkg/apis/security/v1alpha1"
@@ -43,6 +45,8 @@ func TestWCClientCert(t *testing.T) {
 		provider             string
 		capi                 bool
 		clustersInNamespaces map[string]string
+		isAdmin              bool
+		permittedResources   []v1.ResourceRule
 		expectError          *microerror.Error
 	}{
 		// Logging into WC
@@ -54,6 +58,7 @@ func TestWCClientCert(t *testing.T) {
 				WCCertTTL: "8h",
 			},
 			provider: "aws",
+			isAdmin:  true,
 		},
 		// Logging into WC that does not exist
 		{
@@ -64,6 +69,7 @@ func TestWCClientCert(t *testing.T) {
 				WCCertTTL: "8h",
 			},
 			provider:    "aws",
+			isAdmin:     true,
 			expectError: clusterNotFoundError,
 		},
 		// self contained file
@@ -76,6 +82,14 @@ func TestWCClientCert(t *testing.T) {
 				SelfContained: "/cluster.yaml",
 			},
 			provider: "aws",
+			permittedResources: []v1.ResourceRule{
+				{
+					Verbs:         []string{"get"},
+					Resources:     []string{"organizations"},
+					APIGroups:     []string{"security.giantswarm.io/v1alpha1"},
+					ResourceNames: []string{"organization"},
+				},
+			},
 		},
 		// keeping MC context
 		{
@@ -87,6 +101,14 @@ func TestWCClientCert(t *testing.T) {
 				KeepContext: true,
 			},
 			provider: "aws",
+			permittedResources: []v1.ResourceRule{
+				{
+					Verbs:         []string{"get"},
+					Resources:     []string{"organizations"},
+					APIGroups:     []string{"security.giantswarm.io/v1alpha1"},
+					ResourceNames: []string{"organization"},
+				},
+			},
 		},
 		// Explicit organization
 		{
@@ -98,6 +120,7 @@ func TestWCClientCert(t *testing.T) {
 				WCOrganization: "organization",
 			},
 			provider: "aws",
+			isAdmin:  true,
 		},
 		// Several clusters in several namespaces exist
 		{
@@ -108,6 +131,7 @@ func TestWCClientCert(t *testing.T) {
 				WCCertTTL: "8h",
 			},
 			provider: "aws",
+			isAdmin:  true,
 		},
 		// Trying to log into a cluster in default namespace without insecure namespace
 		{
@@ -117,7 +141,15 @@ func TestWCClientCert(t *testing.T) {
 				WCName:    "cluster",
 				WCCertTTL: "8h",
 			},
-			provider:    "aws",
+			provider: "aws",
+			permittedResources: []v1.ResourceRule{
+				{
+					Verbs:         []string{"get"},
+					Resources:     []string{"organizations"},
+					APIGroups:     []string{"security.giantswarm.io/v1alpha1"},
+					ResourceNames: []string{"organization"},
+				},
+			},
 			expectError: clusterNotFoundError,
 		},
 		// Trying to log into a cluster in default namespace with insecure namespace
@@ -130,6 +162,14 @@ func TestWCClientCert(t *testing.T) {
 				WCInsecureNamespace: true,
 			},
 			provider: "aws",
+			permittedResources: []v1.ResourceRule{
+				{
+					Verbs:         []string{"get"},
+					Resources:     []string{"organizations"},
+					APIGroups:     []string{"security.giantswarm.io/v1alpha1"},
+					ResourceNames: []string{"organization"},
+				},
+			},
 		},
 		// Trying to log into a cluster on kvm
 		{
@@ -140,6 +180,7 @@ func TestWCClientCert(t *testing.T) {
 				WCCertTTL: "8h",
 			},
 			provider:    "kvm",
+			isAdmin:     true,
 			expectError: unsupportedProviderError,
 		},
 		// Trying to log into a cluster on azure
@@ -151,6 +192,7 @@ func TestWCClientCert(t *testing.T) {
 				WCCertTTL: "8h",
 			},
 			provider: "azure",
+			isAdmin:  true,
 		},
 		// Trying to log into a cluster on openstack
 		{
@@ -162,6 +204,14 @@ func TestWCClientCert(t *testing.T) {
 			},
 			provider: "openstack",
 			capi:     true,
+			permittedResources: []v1.ResourceRule{
+				{
+					Verbs:         []string{"get"},
+					Resources:     []string{"organizations"},
+					APIGroups:     []string{"security.giantswarm.io/v1alpha1"},
+					ResourceNames: []string{"organization"},
+				},
+			},
 		},
 	}
 
@@ -194,6 +244,9 @@ func TestWCClientCert(t *testing.T) {
 			}
 
 			client := kubeclient.FakeK8sClient()
+			client.AddSubjectAccess(tc.isAdmin)
+			client.AddSubjectResourceRules("default", tc.permittedResources)
+
 			ctx := context.Background()
 			{
 				err = client.CtrlClient().Create(ctx, getOrganization("org-organization"))