+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 - Switched from exp to non-exp apiVersion for `MachinePools` and `AzureMachinePools` CR on `Azure` in `get nodepool` command.
 
+## Added
+
+- Added `organizations` subcommand to `kubectl gs get` family of commands to list and display details of organizations
+
 ## [2.20.0] - 2022-09-02
 
 ## Added